Remove plugin support
