+++ src/main/scala/org/calinburloiu/music/microtuner/MicrotonalistApp.scala
@@ -24,7 +24,7 @@ import com.typesafe.scalalogging.StrictLogging
 import org.calinburloiu.music.intonation.format.{LocalScaleLibrary, ScaleFormatRegistry}
 import org.calinburloiu.music.microtuner.format.JsonScaleListFormat
 import org.calinburloiu.music.microtuner.midi._
-import org.calinburloiu.music.tuning.{Tuning, TuningList, TuningMapperRegistry, TuningReducerRegistry}
+import org.calinburloiu.music.tuning.{Tuning, TuningList}
 
 import scala.util.Try
 
@@ -80,8 +80,7 @@ object MicrotonalistApp extends StrictLogging {
 
     // # I/O
     val scaleLibraryPath = mainConfigManager.coreConfig.scaleLibraryPath
-    val scaleListReader = new JsonScaleListFormat(new LocalScaleLibrary(ScaleFormatRegistry, scaleLibraryPath),
-      new TuningMapperRegistry, new TuningReducerRegistry)
+    val scaleListReader = new JsonScaleListFormat(new LocalScaleLibrary(ScaleFormatRegistry, scaleLibraryPath))
 
     // # Microtuner
     val scaleList = scaleListReader.read(new FileInputStream(inputFileName))
