+++ src/main/scala/org/calinburloiu/music/tuning/DirectTuningReducer.scala
@@ -17,9 +17,8 @@
 package org.calinburloiu.music.tuning
 
 import com.typesafe.scalalogging.StrictLogging
-import org.calinburloiu.music.plugin.{PluginConfig, PluginFactory}
 
-class DirectTuningReducer extends TuningReducer(None) with StrictLogging {
+class DirectTuningReducer extends TuningReducer with StrictLogging {
 
   override def apply(partialTuningList: PartialTuningList): TuningList = {
     val maybeTunings = partialTuningList.tuningModulations.map { modulation =>
@@ -37,25 +36,6 @@ class DirectTuningReducer extends TuningReducer(None) with StrictLogging {
       throw new DirectTuningReducerException(
         s"Some tunings are not complete: $maybeTunings")
     }
-  }
-}
-
-object DirectTuningReducer {
-  val pluginId = "direct"
-}
-
-class DirectTuningListReducerFactory extends PluginFactory[DirectTuningReducer] {
-
-  override val pluginId: String = DirectTuningReducer.pluginId
-
-  override val configClass: Option[Class[_ <: PluginConfig]] = None
-
-  override val defaultConfig: Option[PluginConfig] = None
-
-  override def create(config: Option[PluginConfig] = None): DirectTuningReducer = config match {
-    case None => new DirectTuningReducer
-    case otherConfig => throw new IllegalArgumentException(
-      s"Expecting a specific DirectTuningListReducer, but got ${otherConfig.getClass.getName}")
   }
 }
