+++ src/test/resources/scale_lists/minor_major.scalist
@@ -27,10 +27,8 @@
         "ref": "chromatic.scl"
       },
       "fillTuningMapper": {
-        "id": "auto",
-        "config": {
-          "mapQuarterTonesLow": true
-        }
+        "type": "auto",
+        "mapQuarterTonesLow": true
       }
     },
     {
